Framework is now more modular
Now only jQuery, one ajax call, and codemirror libs are required to get
up and running!

For now, calls to /compile on the server use an external api (rextester)
but will be done on the server vm later on internally. This api is being
used while other features are being worked on.